+++ js/config/index.js
@@ -4,6 +4,9 @@
 module.exports = {
   configs: {
     'js-rules': require('./configs/eslintrc.js'),
+    'import-rules': require('./configs/import-eslintrc.js'),
+    'react-rules': require('./configs/react-eslintrc.js'),
+    'react-with-hooks-rules': require('./configs/react-with-hooks-eslintrc.js'),
     'typescript-rules': require('./configs/typescript-eslintrc.js'),
   },
 }